domain grouped project restructure
