+++ lib/categories/screens/categories_screen.dart
@@ -1,7 +1,9 @@
-import 'package:expense_categoriser/services/categories_provider.dart';
-import 'package:expense_categoriser/utils/models/category.dart';
+import 'package:expense_categoriser/categories/utils/models/category.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import '../providers/categories_provider.dart';
+import '../ui/category_keyword_field.dart';
 
 final showNewCategoryField = StateProvider<bool>((ref) => false);
 final selectedCategoryId = StateProvider<int>((ref) => 0);
@@ -119,7 +121,15 @@ class CategoryDetails extends ConsumerWidget {
                         .read(categoriesProvider.notifier)
                         .deleteCategory(category)),
                 Text(category.name),
-                CategoryKeywordField(category: category),
+                CategoryKeywordField(
+                  category: category,
+                  onChange: (String addedCategory) {
+                    category.keywords = [...category.keywords, addedCategory];
+                    ref
+                        .read(categoriesProvider.notifier)
+                        .updateCategory(category);
+                  },
+                ),
                 for (var keyword in category.keywords)
                   MaterialButton(
                       child: Text(keyword),
@@ -134,49 +144,5 @@ class CategoryDetails extends ConsumerWidget {
           }
           return Container();
         });
-  }
-}
-
-class CategoryKeywordField extends ConsumerStatefulWidget {
-  final Category category;
-  const CategoryKeywordField({Key? key, required this.category})
-      : super(key: key);
-
-  @override
-  ConsumerState<CategoryKeywordField> createState() =>
-      CategoryKeywordFieldState();
-}
-
-class CategoryKeywordFieldState extends ConsumerState<CategoryKeywordField> {
-  TextEditingController controller = TextEditingController();
-
-  @override
-  void initState() {
-    super.initState();
-    controller.addListener(_handleChanges);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        SizedBox(
-          width: 240,
-          child: TextField(
-            controller: controller,
-          ),
-        ),
-      ],
-    );
-  }
-
-  _handleChanges() {
-    final text = controller.text;
-    if (text.isNotEmpty && text.substring(text.length - 1) == ',') {
-      final addedCategory = text.substring(0, text.length - 1);
-      widget.category.keywords = [...widget.category.keywords, addedCategory];
-      ref.read(categoriesProvider.notifier).updateCategory(widget.category);
-      controller.clear();
-    }
   }
 }
