+++ lib/categories/services/categoriser.dart
@@ -1,7 +1,8 @@
-import '../utils/enums/numbering_style.dart';
-import '../utils/models/category.dart';
-import '../utils/models/import_settings.dart';
-import '../utils/models/transaction.dart';
+import 'package:expense_categoriser/categories/utils/enums/numbering_style.dart';
+import 'package:expense_categoriser/categories/utils/models/category.dart';
+import 'package:expense_categoriser/csv_files/utils/models/import_settings.dart';
+
+import '../../reports/utils/models/transaction.dart';
 
 class Categoriser {
   List<Category> categories;
