+++ lib/reports/screens/report_screen.dart
@@ -1,18 +1,16 @@
 import 'dart:async';
 
-import 'package:expense_categoriser/reports/models/report.dart';
+import 'package:expense_categoriser/categories/utils/models/category.dart';
+import 'package:expense_categoriser/csv_files/providers/csv_files_provider.dart';
+import 'package:expense_categoriser/reports/utils/models/report.dart';
 import 'package:expense_categoriser/reports/view_models/report_view_model.dart';
-import 'package:expense_categoriser/services/categoriser.dart';
-import 'package:expense_categoriser/services/csv_reader_service.dart';
-import 'package:expense_categoriser/reports/view_models/report_service.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../services/categories_provider.dart';
-import '../../services/csv_files_provider.dart';
+import '../../categories/providers/categories_provider.dart';
+import '../../csv_files/services/csv_reader_service.dart';
 import '../ui/uncategorised_item_row.dart';
-import '../../utils/models/category.dart';
-import '../../utils/models/transaction.dart';
+import '../utils/models/transaction.dart';
 
 class ReportScreen extends ConsumerStatefulWidget {
   const ReportScreen({super.key});
@@ -22,7 +20,6 @@ class ReportScreen extends ConsumerStatefulWidget {
 }
 
 class _ReportScreenState extends ConsumerState<ReportScreen> {
-  final ReportService reportService = ReportService();
   final CsvReaderService csvReaderService = CsvReaderService();
 
   @override
