+++ lib/reports/ui/selectable_words.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../models/selectable_word_item.dart';
+import '../utils/models/selectable_word_item.dart';
 
 class SelectableWords extends StatefulWidget {
   const SelectableWords({
