+++ lib/reports/utils/models/report.dart
@@ -1,4 +1,4 @@
-import '../../utils/models/transaction.dart';
+import 'transaction.dart';
 
 class ReportCategory {
   num total = 0;
