+++ lib/shared/ui/app_menu.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../utils/constants/pages.dart';
+import '../pages.dart';
 
 final selectedPageNameProvider = StateProvider<String>((ref) {
   return pages.keys.first;
